requirements.txt
asgiref==3.5.2
Django==4.0.4
django-ckeditor==6.4.1
django-etc==1.3.1
django-js-asset==2.0.0
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
